move session manager javascript file
