+++ web/src/static/js/managers/session-manager.js
