+++ web/src/static/js/vote-api/voteUtils.js
@@ -2,7 +2,7 @@
  * Vote API utilities for handling upvotes and downvotes
  */
 
-import { sessionManager } from "../auth/session-manager-vanilla.js";
+import { sessionManager } from "../managers/session-manager.js";
 
 const API_BASE_URL = 'http://localhost:8000'; // Adjust this to your API URL
